Glassfish 4.1.1 Support verified.
Addressed issue with collection types. It works under Glassfish because
the Jersey runtime is constructing the response object. When using
WebSphere Liberty, it is still failing because the Apache CXF runtime is
somehow being invoked only in providing the implementation of the
response and the Jersey code does not properly copy the contents of the
generic types.